+++ src/web3py/middleware.py
@@ -2,12 +2,11 @@ import logging
 from typing import Any, Callable
 from urllib.parse import urlparse
 
-from requests import Response, HTTPError
+from requests import HTTPError, Response
 from web3 import Web3
 from web3.types import RPCEndpoint, RPCResponse
 
-from src.metrics.prometheus.basic import ETH1_RPC_REQUESTS_DURATION, ETH1_RPC_REQUESTS
-
+from src.metrics.prometheus.basic import ETH1_RPC_REQUESTS, ETH1_RPC_REQUESTS_DURATION
 
 logger = logging.getLogger(__name__)
 
@@ -26,7 +25,7 @@ def metrics_collector(
     def middleware(method: RPCEndpoint, params: Any) -> RPCResponse:
         try:
             # Works only with HTTP and Websocket Provider
-            domain = urlparse(w3.provider.endpoint_uri).netloc
+            domain = urlparse(getattr(w3.provider, "endpoint_uri")).netloc
         except:
             domain = 'unavailable'